% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{combine_freq_sev_ci}
\alias{combine_freq_sev_ci}
\title{combine_freq_sev_ci}
\usage{
combine_freq_sev_ci(
  freq_sev_ci,
  peril_names,
  peril_distributions,
  percentile_granularity = 100,
  return_percentiles = c(0.025, 0.5, 0.975),
  percentile_interval = c(0.5, 0.99)
)
}
\arguments{
\item{freq_sev_ci}{Input list of all peril names and a list of the distributions attributable to those perils.}

\item{peril_names}{List of all the peril names.}

\item{peril_distributions}{List of the distributions attributable to the perils.}

\item{percentile_granularity}{Set to 100.}

\item{return_percentiles}{0.025, 0.5, 0.975.}

\item{percentile_interval}{0.5, 0.99.}
}
\description{
combine_freq_sev_ci
}
\examples{
return_val <- combine_freq_sev_ci(freq_sev_ci = fvci,
                                  peril_names = all_perils,
                                  peril_distributions = peril_dists)
}
