% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{dBeta_ab}
\alias{dBeta_ab}
\title{dBeta_ab}
\usage{
dBeta_ab(x, shape1 = 2, shape2 = 3, b = 1)
}
\arguments{
\item{x}{A vector.}

\item{shape1}{Shape parameter.}

\item{shape2}{Shape parameter.}

\item{b}{Boundary parameter.}
}
\value{
Density function.
}
\description{
Density function for the 4-parameter Beta distribution.
}
