% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{dgumbel}
\alias{dgumbel}
\title{dgumbel}
\usage{
dgumbel(x, mu, s)
}
\arguments{
\item{x}{Vector.}

\item{mu}{Location parameter.}

\item{s}{Scale parameter.}
}
\value{
Density Function.
}
\description{
Density function for the Gumbel distribution with location and scale parameters.
}
\examples{
dgumbel(-1:2, -1, 0.5)
}
