% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_fitting.R
\name{filter_distribution}
\alias{filter_distribution}
\title{Filter distribution}
\usage{
filter_distribution(fitted_distribution = NULL)
}
\arguments{
\item{fitted_distribution}{5 column data frame produced by the fit_distribution function.}
}
\value{
Filtered data frame with the distribution with the smallest AIC score for each peril.
}
\description{
This function finds the distribution with the smallest aic score from fitted_distribution.
}
\examples{
fd <- fitted_distribution()[[1]]
filter_distribution(fd)
}
