% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_fitting.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Fit distribution}
\usage{
fit_distribution(the_right_data = NULL, updateProgress, data_type, years)
}
\arguments{
\item{updateProgress}{}
}
\value{
Data frame with all the distribution fits for each peril.
}
\description{
This function takes the data, after all user inputs and checks to see if the data is ready to be fit. If the data is ready to be fit, then it uses the get_aic_mle function on the data.
}
\examples{
rd <- get_right_data()
rd <- rd[[1]]
temp <- fit_distribution(rd, updateProgress)
}
