% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_fitting.R
\name{flexible_samples}
\alias{flexible_samples}
\title{Flexible samples}
\usage{
flexible_samples(mle_list_input, dist_name, n)
}
\arguments{
\item{n}{}
}
\description{
This function is flexible for function name and samples based on input mle values. Used within the get_freq_sev_ci function.
}
\examples{

return_sample <- flexible_samples(n=number_of_simulations,
                                  dist_name = paste0(dist),
                                  mle_list_input= lapply(ci_mle_list[[peril]][[paste0(dist)]],"[",j))
                flexible_samples(n=number_of_simulations,
                                 dist_name = "bern",
                                 mle_list_input=freq_boot[j])
}
