% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_fitting.R
\name{get_aic}
\alias{get_aic}
\title{Get AIC}
\usage{
get_aic(llhood, number_of_parameters)
}
\arguments{
\item{llhood}{Log-likelihood}

\item{number_of_parameters}{The number of parameters in the distribution}
}
\value{
A number.
}
\description{
This function computes the AIC score from the log likelihood for distributions that don't have built in R functions.
}
