% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{get_freq_sev_ci}
\alias{get_freq_sev_ci}
\title{get_freq_sev_ci}
\usage{
get_freq_sev_ci(
  bootstraps,
  number_of_simulations = 1000,
  number_of_resamples = 1000,
  percentile_granularity = 100,
  return_percentiles = c(0.025, 0.5, 0.975),
  percentile_interval = c(0.5, 0.99)
)
}
\arguments{
\item{number_of_simulations}{Number of simulations, default is 1000}

\item{number_of_resamples}{Number of resamples, default is 1000}

\item{the_right_data}{Data frame.}

\item{ci_mle_list}{List of the chosen MLE parameters for each peril.}
}
\description{
This function takes the MLEs from the previous outputs and uses the input data to calculate a frequency distribution and hence estimates a freq*sev distribution for each combination of perils
}
\examples{
return_list <- get_freq_sev_ci(the_right_data = input_data, ci_mle_list = distribution_mles)
}
