% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_functions.R
\name{get_gap_curve}
\alias{get_gap_curve}
\title{get_gap_curve}
\usage{
get_gap_curve(sim_vector, budget = budget, is_ci = FALSE)
}
\arguments{
\item{sim_vector}{Vector of simulated values for the selected peril.}

\item{budget}{Selected budget}

\item{is_ci}{If user has selected to show confidence intervals, default = FALSE.}
}
\value{
Data frame for funding gap curve.
}
\description{
A function that creates the funding gap curve for Output 4
}
\examples{
# Create funding gap curve
curve <- get_gap_curve(dat_sim$value, budget = budget)

if (!is.null(perils_ci) & !any(is.na(perils_ci))) {
  number_of_quantiles <- length(perils_ci$average)

  curve_lower <- get_gap_curve(perils_ci$percentiles[1,paste0(seq(0.5,0.98,by=0.01)*100,"\%")],
                               budget = budget,
                               is_ci = TRUE)

  curve_upper <- get_gap_curve(perils_ci$percentiles[number_of_quantiles,paste0(seq(0.5,0.98,by=0.01)*100,"\%")],
                               budget = budget,
                               is_ci = TRUE)

  curve$value_lower <- curve_lower$`Funding gap`
  curve$value_upper <- curve_upper$`Funding gap`
} else {
  curve$value_lower <- NA
  curve$value_upper <- NA
}
}
