% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_functions.R
\name{plot_simulations}
\alias{plot_simulations}
\title{plot_simulations}
\usage{
plot_simulations(
  rs = NULL,
  right_data = NULL,
  chosen_peril = NULL,
  overlap = NULL
)
}
\arguments{
\item{rs}{Simulated data from the run_simulations function.}

\item{right_data}{Data frame, which is either the scaled, detrended or core data, depending on the inputs selected}

\item{overlap}{}

\item{peril}{Selected peril type}
}
\description{
This function takes the simulated data and the original loss data and plots two overlapping densities, comparing the distribution of the loss data with the simulated data for each peril.
}
\examples{
 plot_simulations(rs = rs,
 right_data = rd,
 peril = ips,
 overlap = ioc) + xlab(paste0('value ',currency_code())) +
   coord_cartesian(xlim = c(0, max(observed_data$value * 1.25))) +
   scale_fill_manual(values=fill_colours) +
   theme(legend.title=element_blank())
}
