% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{prepare_simulations}
\alias{prepare_simulations}
\title{prepare_simulationa}
\usage{
prepare_simulations(
  fitted_distribution = NULL,
  dist_flood = NULL,
  dist_drought = NULL,
  dist_storm = NULL,
  dist_earthquake = NULL
)
}
\arguments{
\item{fitted_distribution}{5 column data frame as produced by the fit_distribution function.}

\item{dist_flood}{Selected distribution for flood.}

\item{dist_drought}{Selected distribution for drought.}

\item{dist_storm}{Selected distribution for storm.}

\item{dist_earthquake}{Selected distribution for earthquake.}
}
\value{
Data frame.
}
\description{
This function prepares the data for simulations
}
\examples{
prepared_simulations <- reactive({
fd <- fitted_distribution()[[1]]
x <- prepare_simulations(fd, dist_flood = input$dist_flood_input,
                         dist_drought = input$dist_drought_input,
                         dist_storm = input$dist_storm_input,
                         dist_earthquake = input$dist_earthquake_input)
x
})
}
