% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{rfrechet}
\alias{rfrechet}
\title{rfrechet}
\usage{
rfrechet(n, alpha, s)
}
\arguments{
\item{n}{Number of observations.}

\item{alpha}{Shape parameter.}

\item{s}{Scale parameter.}
}
\value{
Binary vector of length \code{n}.
}
\description{
Generate random deviates from the Frechet distribution with shape and scale parameters.
}
\examples{
rfrechet(6, 0.8, 0.5)
}
