% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{sample_skewness}
\alias{sample_skewness}
\title{Sample Skewness}
\usage{
sample_skewness(x)
}
\arguments{
\item{x}{Sample data.}
}
\value{
A number.
}
\description{
This function calculates the adjusted Fisher-Pearson coefficient of skewness, adjusted for sample size.
}
\examples{
sample_skew <- sample_skewness(data)
}
