% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_functions.R
\name{sample_var}
\alias{sample_var}
\title{Sample Variance}
\usage{
sample_var(x)
}
\arguments{
\item{x}{Sample data.}
}
\value{
A number.
}
\description{
This function calculates the sample variance.
}
\examples{
sample_var <- sample_var(data)
}
