% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_detrending.R
\name{test_linear_trend}
\alias{test_linear_trend}
\title{test_linear_trend}
\usage{
test_linear_trend(dat)
}
\arguments{
\item{dat}{Data frame}
}
\value{
Data frame with the perils and their p-values.
}
\description{
This function tests through each peril if a linear trend exists in the data.
}
\examples{
test_data <- sdr
all_perils <- test_linear_trend(test_data)
}
